requirements.txt
Flask>=2.0
qrcode[pil]
Pillow
python-barcode
